all might: FINAL VERSION PLEASE . Ánh

+++ src/java/Control/DeleteProducts.java
@@ -6,6 +6,7 @@
 package Control;
 
 import BEAN.Product;
+import BEAN.User;
 import DAO.ProductDAO;
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -14,6 +15,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 /**
  *
@@ -33,13 +35,19 @@ public class DeleteProducts extends HttpServlet {
      */
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        int id = Integer.parseInt(request.getParameter("id"));
-        Product product = new Product();
-        product.setId(id);
-        ProductDAO productDAO = new ProductDAO();
-        productDAO.deleteProduct(id);
-        response.sendRedirect("/FlowersShop/nManager");
+        HttpSession session = request.getSession();
+        if (session.getAttribute("user") != null) {
+            User user = (User) session.getAttribute("user");
+            if (user.getRole() == 0) {
+                response.setContentType("text/html;charset=UTF-8");
+                int id = Integer.parseInt(request.getParameter("id"));
+                Product product = new Product();
+                product.setId(id);
+                ProductDAO productDAO = new ProductDAO();
+                productDAO.deleteProduct(id);
+                response.sendRedirect("/FlowersShop/nManager");
+            }
+        }
     }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
